fixes: bug fixes of disk and fat

+++ src/bootloader/stage2/c/external/x86.h
@@ -9,5 +9,5 @@ bool _cdecl x86_disk_reset(u8 drive);
 bool _cdecl x86_disk_read(u8 drive, u16 cylinder, u16 sector, u16 head,
                           u8 count, void far *out_data);
 bool _cdecl x86_disk_get_drive_params(u8 drive, u8 *out_drive_type,
-                                      u8 *out_cylinders, u16 *out_sectors,
+                                      u16 *out_cylinders, u16 *out_sectors,
                                       u16 *out_heads);
